initscripts: rebase on latest centos
Summary:
This is a full rebase on the version of initscripts in 7.4 which is itself a clean rebase on master. As such we don't need most of our patches anymore. It's also why the Changelog is quite a bit different.

I did pull in 3 patches we need that are not upstream:
* IFDOWN_ON_SHUTDOWN support
* 2 patches for EXTERNAL tunnel support

Reviewed By: davide125

Differential Revision: D7513521

fbshipit-source-id: 4f86832282e5d08dfb762f847b34d29b8414cb22