Bump release to fb3, no changes
Summary:
Back in 2019, the fb2 was published to the repos, but was never commited, and never updated in slowroll. TW uses latest versions of the packages these days, so, when we removed the fb2 (because of conflict in ROU), their image builds started failing, because of downgrade conflict on chef run, as chef respects slowroll and they do not.
Build fb3, exactly the same as fb1 and publish it properly (I will roll it out with slowroll later), so, shit stops breaking.

Reviewed By: malmond77

Differential Revision: D20871590

fbshipit-source-id: c71b653955748be44e58e6bc05853929b309806a